Remove AstraeaUI.java
The AstraeaUI contains methods used for printing formatted messages
to console.

As the JavaFX GUI was introduced in previous commits, the console
UI is no longer required.

Let's remove AstraeaUI and all its usages to reduce code duplication
and complexity.

The usage of ExitCommand from "bye" will also need to be
reconsidered, since the GUI window can be closed to
terminate the program.